+++ src/app/event/[id].tsx
@@ -2,8 +2,8 @@
 
 import { useRouter } from 'next/router'
 import type { ReactNode, FC } from 'react'
-import { EventDetails, type EventDetailType } from '../../components/EventList'
-import { Layout } from '../../components/Layout'
+import { EventDetails, type EventDetailType } from '../../src/components/EventList'
+import { Layout } from '../../src/components/Layout'
 
 const defaultId = 'rupc2018'
 const getEntries: Function = (contentId: string, key: keyof EventDetailType) => {
