+++ src/app/layout.tsx
@@ -1,0 +1,21 @@
+import type { Metadata } from 'next'
+import type { ReactNode } from 'react'
+import './globals.css'
+
+export const metadata: Metadata = {
+  title: 'RiPPro(立命館大学情報理工学部プロジェクト団体)',
+  description:
+    '立命館大学情報理工学部プロジェクト団体 RiPPro (Ritsumeikan Programming Project) は、ACM-ICPC (プログラミングの学生世界大会) などの大会で入賞することを目的として結成されました。'
+}
+
+type RootLayoutProps = {
+  children: ReactNode
+}
+
+export default function RootLayout({ children }: RootLayoutProps) {
+  return (
+    <html lang="ja">
+      <body>{children}</body>
+    </html>
+  )
+}
